test: remove reliance on persistencetestsupport
For unknown reasons, Xcode 14 is unable to build the previous the
project for testing on the previous commit, complaining about the
module Dependencies being missing while building Persistence. The
Persistence module can be build on its own, but when building the
whole macos test scheme, it seems to be missing (possible due to
incorrect build order?).

To get around this issue, remove reliance on PersistenceTestSupport
from the macOS app test target.