fix(profiles): optimise fetching server groups
Include server group information in the ModelState struct. We don't
want to re-fetch server groups multiple times per state change, since
it's enough to do it just once whenever the server type is changed.

This also prevents groups being fetched with incorrect features and
countries specified, which was triggering an assertion failure and
resetting the selected server/country to a different one.

Jira-Id: VPNAPPL-2075