fix: use forked credentials if needed on token refresh
Previously, if `handleTokenExpired` was used to get a new set of
credentials as part of consuming a forked session selector, and that
invocation failed because of a timeout or some other error, then the
partial credentials returned from consuming the selector would not be
re-used on the second try, resulting in a failure to renew the session
until the application tried a second time to refresh the certificate.
This is a difficult condition to reproduce, but if the request times
out, connectivity can be impacted until the user interacts with the
application again, which can take hours.