chore: fix PMLogger helpers causing building issues with Xcode 15.3
Xcode 15.3 seems to cause errors in place where it was building fine
with older Xcode releases. Errors are of the following type:

  > Reference to property 'taskId' in closure requires explicit use of
  > 'self' to make capture semantics explicit

This is triggerred especially with logging messages passing metadata
dictionaries via the `metadata` parameter which themselves might
reference instances variables and thus needing explicit capture
semantics.

This changes the API of the PMLogger helpers to make the metadata
parameter in those helpers non escaping anymore (since they never really
outlive the function call itself) but it is using internally the
`withoutActuallyEscaping` of the Swift StdLib since the implementation
of those helpers rely on a function needing an escaping closure.