feat(connection): extension management
This commit introduces a new Connection package, superceding the old
LegacyCommon.VpnManager, as the latter proved to be too difficult to
extract due to heavy coupling and risk of breaking functionality in
the iOS and MacOS apps.

The new Connection package aims to modularise connection logic.
LocalAgent and CertificateAuthentication functionality is stubbed out
for now.

Jira-Id: VPNAPPL-2170