fix(connection): failure after logging out
LocalAgentFeature was not receiving actions whenever events were
emitted by LocalAgentClientImplementation due to an error in the way in
which we subscribed to these events. This happened after logging out
and logging back in. This commit ensures that a new AsyncStream is
created each time, so that events are received even after logging out
and back in.