fix(connection): prevent infinite disconnection
Cancel the timeout started on connection attemps when whenever the
connection feature fully disconnects, whether this is user initiated
or not. This prevents us from being stuck in the disconnecting state
if e.g. we receive a local agent error and a disconnect is started
before we are able to connect.