fix(connection): prevent infinite disconnection
If the extension manager feature receives the command to disconnect
before the tunnel has been started, no .disconnected event will be
received by the tunnel since the tunnel is already disconnected.

This commit adds a new .preparingConnection state which represents the
managers and configurations being prepared, before the tunnel is
started.